+++ dbmodel/schemas/main/ws/updates/4/12/0/patch.sql
@@ -388,3 +388,11 @@ UPDATE config_toolbox
   }
 ]'::json
 	WHERE id=3110;
+
+-- Add column
+ALTER TABLE man_tank ADD COLUMN vnom numeric (12,3);
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'vnom', 'lyt_data_1', 54, 'double', 'text', 'vnom', NULL, NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature WHERE feature_class = 'TANK'
+ON CONFLICT DO NOTHING;